refactor : service.js와 controller.js 파일 병합
